requirements.txt
runpod>=1.5.1
torch>=2.0.0
torchvision
torchaudio
diffusers[torch]>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
safetensors>=0.3.0
Pillow>=9.0.0
numpy>=1.24.0
xformers
compel
requests
huggingface-hub>=0.16.0
